requirements.txt
librosa==0.6.2
numpy==1.16.5
Pillow==6.2.0
pydub==0.23.0
scipy==1.3.1
tensorflow==2.0.0